First part copy object encryption fixes.
This is most of the first phase of the copy object encryption fixes.
It comprises additional data paths to communicate data from the
rest layer down to the sal layer to handle decrypting
objects.

Past this, a future set of commits based on this needs to use the same
data path to pass down re-encrypt data for the destination.

Signed-off-by: Marcus Watts <mwatts@redhat.com>