qa: use distinct names for CephFSs created during testing
Right now, our test code creates a CephFS with the name "cephfs". This
name is too generic and that makes it impossible to track it in the
logs.

We use the name "cephfs" quite often -

* a CephFS named "cephfs"
* a python module named "cephfs" (that contains functional tests for CephFS)
* a config file section named "cephfs"
* package names starting with or ending with "cephfs"
* CephFS related tools start with string "cephfs" (example: cepfs-shell,
  ceph-top, cephfs-data-scan, etc.)
* daemon names start with "cephfs" (example: cephfs-mirror)
* OSD pool application tag named "cephfs"

So a unique/distinct name for CephFS will make it easy to go track it in
logs.

CephFSs are freshly created for each test method/test case (after
destroying the CephFSs from previous test method/test case). Use new and
distinct for each test method. This will make tracking a CephFS for a
specifc test method even more easier.

Fixes: https://tracker.ceph.com/issues/62876
Signed-off-by: Rishabh Dave <ridave@redhat.com>