mgr/snap_schedule: handle volume delete
Remove references to timers and db connection that refer to paths in the
deleted volume.

This is a best case effort and not 100% safe. There can be race
conditions where the db connection is closed while the module is
executing a query on the db while the volume itself is deleted.

Hopefully there won't be many tracebacks seen.

Fixes: https://tracker.ceph.com/issues/63999
Signed-off-by: Milind Changire <mchangir@redhat.com>