osd: roll_forward_to->pg_committed_to in EC*
The purpose of this rename is merely to clarify that the necessary
condition on ec roll-forward is that the pg has committed up to that
point.  Along with subsequent commits, this will clarify that both ec
and replicated pools propagate pg_committed_to for related if not
identical reasons.  Because EC::submit_transaction already did

  op->roll_forward_to = std::max(min_last_complete_ondisk, rmw_pipeline.committed_to);

there's no difference in behavior as rmw_pipeline.committed_to is
updated immediately after the notification to the PG that the write
completed.

Signed-off-by: Samuel Just <sjust@redhat.com>