rgw: forward x-amz-content-sha256 to master if body not empty
For example, when forwarding a bucket creation request to the master,
the request body is empty. However, the user might have provided an
x-amz-content-sha256 for a request that includes a
CreateBucketConfiguration document in the body. Since the request is
being forwarded without a body, this will cause authentication to fail
on the master if the user-provided x-amz-content-sha256 is used due to
the signature mismatch.

Signed-off-by: Seena Fallah <seenafallah@gmail.com>